upd: remove #include 'main.h' from wallet.h
