+++ divi/src/main.h
@@ -343,13 +343,7 @@ void PopulateInvalidOutPointMap();
 bool ValidOutPoint(const COutPoint out, int nHeight);
 bool RecalculateDIVSupply(int nHeightStart);
 
-/**
- * Check if transaction will be final in the next block to be created.
- *
- * Calls IsFinalTx() with current block height and appropriate block time.
- *
- * See consensus/consensus.h for flag definitions.
- */
+
 bool CheckFinalTx(const CTransaction& tx, int flags = -1);
 
 /** Check for standard transaction types
@@ -357,7 +351,7 @@ bool CheckFinalTx(const CTransaction& tx, int flags = -1);
  */
 bool IsStandardTx(const CTransaction& tx, std::string& reason);
 
-bool IsFinalTx(const CTransaction& tx, int nBlockHeight = 0, int64_t nBlockTime = 0);
+// TODO: delete : bool IsFinalTx(const CTransaction& tx, int nBlockHeight = 0, int64_t nBlockTime = 0);
 
 /** Undo information for a CBlock */
 class CBlockUndo
