+++ divi/src/test/rpc_wallet_tests.cpp
@@ -7,7 +7,7 @@
 
 #include "base58.h"
 #include "wallet.h"
-
+#include "main.h"
 #include <boost/algorithm/string.hpp>
 #include <boost/test/unit_test.hpp>
 #include "test_only.h"
