+++ divi/src/wallet.h
@@ -8,12 +8,15 @@
 #ifndef BITCOIN_WALLET_H
 #define BITCOIN_WALLET_H
 
+#ifndef BITCOIN_UTILSTRENCODINGS_H
+#include "utilstrencodings.h" //for atoi64
+#endif // BITCOIN_UTILSTRENCODINGS_H
+
 #include "amount.h"
 #include "base58.h"
 #include "crypter.h"
 #include "key.h"
 #include "keystore.h"
-#include "main.h"
 #include "primitives/block.h"
 #include "primitives/transaction.h"
 #include "ui_interface.h"
@@ -22,7 +25,6 @@
 #include "wallet_ismine.h"
 #include "walletdb.h"
 #include "FeeRate.h"
-
 #include <algorithm>
 #include <map>
 #include <set>
@@ -31,6 +33,7 @@
 #include <string>
 #include <utility>
 #include <vector>
+#include <boost/foreach.hpp>
 
 /**
  * Settings
@@ -65,6 +68,19 @@ class CReserveKey;
 class CScript;
 class CWalletTx;
 class CHDChain;
+class CTxMemPool;
+
+bool MoneyRange(CAmount nValueOut);
+
+/**
+ * Check if transaction will be final in the next block to be created.
+ *
+ * Calls IsFinalTx() with current block height and appropriate block time.
+ *
+ * See consensus/consensus.h for flag definitions.
+ */
+bool CheckFinalTx(const CTransaction& tx, int flags);
+bool IsFinalTx(const CTransaction& tx, int nBlockHeight=0, int64_t nBlockTime=0 );
 
 /** (client) version numbers for particular wallet features */
 enum WalletFeature {
@@ -283,7 +299,7 @@ public:
         fCombineDust = false;
         nAutoCombineThreshold = 0;
     }
-    
+
     bool isMultiSendEnabled()
     {
         return fMultiSendMasternodeReward || fMultiSendStake;
