+++ divi/src/walletdb.cpp
@@ -14,7 +14,7 @@
 #include "util.h"
 #include "utiltime.h"
 #include "wallet.h"
-
+#include "main.h"
 #include <boost/filesystem.hpp>
 #include <boost/foreach.hpp>
 #include <boost/scoped_ptr.hpp>
